// HackerNewsWeb/src/app/shared/models/comment.model.ts
export default class Comment {
    id: string = '';
    text: string = '';
    escomment: string = '';
    points: number = 0;
    user: string = '';
    contribution: string = '';
    comment_id: string = '';
    created_at: string = '';
    author:string = '';
    sonComments:Comment[] = [];



    constructor() { 

    }

}